% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{add_derep_names}
\alias{add_derep_names}
\alias{add_derep_names.derep}
\alias{add_derep_names.list}
\title{Add sequence names to a derep object}
\usage{
add_derep_names(derep, ...)

\method{add_derep_names}{derep}(derep, filename, ...)

\method{add_derep_names}{list}(derep, filenames = names(derep), ...)
}
\arguments{
\item{derep}{(object of class \code{\link[dada2:derep-class]{derep}} or a
\code{list} of  such objects) object(s) to add names to.}

\item{...}{passed to methods}

\item{filename}{(\code{character}) name of fasta/q file that the
\code{\link[dada2:derep-class]{derep}} object was derived from.}

\item{filenames}{(\code{character}) name(s) of file(s) that a \code{list} of
\code{\link[dada2:derep-class]{derep}} was derived from.}
}
\value{
(object of class \code{derep}, or a list of such objects) a
    shallow copy of \code{derep}, with an additional member
    "\code{$names}", giving the identifiers for the sequences from the
    original fasta/q file.
}
\description{
Add sequence names to a derep object
}
