% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{combine_bigmaps}
\alias{combine_bigmaps}
\title{Combine raw and denoised reads from multiple regions of the same sequences.}
\usage{
combine_bigmaps(dadamap, rawdata)
}
\arguments{
\item{dadamap}{(\code{\link{dadamap}} object)}

\item{rawdata}{(\code{\link[tibble]{tibble}}) as returned by
\code{\link{summarize_sread}}}
}
\value{
a \code{\link[tibble]{tibble}} giving the sequences for each region
        for each read (uniquely identified by all other columns)
}
\description{
Combine raw and denoised reads from multiple regions of the same sequences.
}
\details{
Both of the inputs should be annotated with identifying columns to
uniquely identify the source \code{\link[dada2:derep-class]{derep}} objects;
this should happen automatically if \code{\link[dada2]{derepFastq}} and
\code{\link[ShortRead]{readFastq}} are called on a list of filenames (there
will be a column "name" in the outputs of \code{\link{dadamap}} and
\code{\link{summarize_sread}}).
}
