% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{dadamap}
\alias{dadamap}
\title{Produce a map between denoised sequences and their original identifiers.}
\usage{
dadamap(derep, dada, filename = NULL, seq_id = NULL, ...)
}
\arguments{
\item{derep}{a \code{\link[dada2]{derep-class}} object or list of such
objects}

\item{dada}{(\link[dada2]{dada-class} object or list of such objects) the
results of a call to \code{\link[dada2]{dada}} on \code{"derep"}}

\item{filename}{(\code{character} of same length as \code{"derep"}) fasta/q
file that is the source for \code{"derep"}, used for setting
unique sequence IDs.  If \code{"derep"} is a named \code{list}, then
the default is to assume that these filenames are the names of the
source files (as returned by \code{\link[dada2]{derepFastq}} when multiple files are
given). To avoid this behavior, pass an explicit \code{NULL}.}

\item{seq_id}{(\code{character} or list of \code{character}) unique sequence
identifiers for the sequences in \code{"derep"}.  These can
alternatively be provided using argument \code{"filename"} or
calling \code{\link{add_derep_names}} on \code{"derep"}.}

\item{...}{additional columns to add to the output.  Names included in the
output by default should be avoided.}
}
\value{
a \code{\link[tibble]{tibble}} with columns:
  \describe{
    \item{\code{$name} (\code{character})}{names from \code{derep}, if it is
    a named list of \code{\link[=add_derep_names]{named_derep}} objects.
    Otherwise absent unless provided in \code{...}.}
    \item{\code{...} (\code{character})}{any additional arguments, passed on
      to \code{\link[tibble]{tibble}}}
    \item{\code{$seq_id} (\code{character})}{the sequence identifiers from
      the original fasta/q file.}
    \item{\code{$derep_idx} (\code{integer})}{the index of each sequence in
      \code{derep$uniques}.}
    \item{\code{$derep_seq} (\code{character})}{the sequences.}
    \item{\code{$dada_seq} (\code{character})}{the denoised sequences.}
}
}
\description{
Produce a map between denoised sequences and their original identifiers.
}
\details{
Columns \code{$derep_seq} and \code{$dada_seq} contain one sequence
per read as plain character strings. Keeping a separate, dereplicated list of
sequences and storing references to them may seem like it would be more
memory efficient, but it is not necessary to do this explicitly, because this
is what R already does with \code{character} vectors; only one copy of each
unique string is actually kept in memory, and everything else is pointers.
}
