% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{find_region_chimeras}
\alias{find_region_chimeras}
\title{Check for bimeras in a subset of regions}
\usage{
find_region_chimeras(region_table, chimset, sample_column, ...)
}
\arguments{
\item{region_table}{(\code{data.frame} containing all the regions in
\code{chimset})}

\item{chimset}{(\code{character} vector, usually of length 3) regions to
concatenate and then check for bimeras.}

\item{sample_column}{(\code{character} scalar) extra column in region_table
which identifies different samples; if present, use
\code{\link[dada2]{isBimeraDenovoTable}} instead of
\code{\link[dada2]{isBimeraDenovo}}.}

\item{...}{passed on to \code{\link[dada2]{isBimeraDenovoTable}} or
\code{\link[dada2]{isBimeraDenovo}}.}
}
\value{
an \code{integer} vector giving the row numbers of
       \code{region_table} which were detected as bimeras.
}
\description{
Check for bimeras in a subset of regions
}
