% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{map_to_best_asv}
\alias{map_to_best_asv}
\title{Replace unmapped raw reads with the nearest ASV}
\usage{
map_to_best_asv(asvs, raw, maxdist = 10)
}
\arguments{
\item{asvs}{(\code{character} vector) ASV sequences mapped to a set of reads.
Should be \code{\link{NA_character_}} for reads which did not map to an
ASV.}

\item{raw}{(\code{character} vector) Raw read sequences for the same set of
reads as \code{asvs}.  May be \code{\link{NA_character_}}}

\item{maxdist}{(\code{numeric} scalar) Maximum Levenshtein distance between
a raw read and an ASV for the read to be mapped to the ASV.}
}
\value{
a \code{character} vector the same length as \code{asvs}, which has
   the closest ASV for each read.
}
\description{
Replace unmapped raw reads with the nearest ASV
}
\details{
The value for element \code{i} of the result is determined as follows:
  \enumerate{
    \item{\code{asvs[i]} is non-\code{NA}: the value from \code{asvs} is used.}
    \item{\code{asvs[i]} and \code{raw[i]} are both \code{NA}: \code{\link{NA_character_}}}
    \item{\code{asvs[i]} is \code{NA}, \code{raw[i]} is non-\code{NA}: \enumerate{
      \item{\code{raw[i]} is less than \code{maxdist} in edit distance from at
      least one of the non-\code{NA} sequences in \code{asvs}: the value from \code{asvs} which has the smallest edit distance from \code{raw[i]} is used}
      \item{\code{raw[i]} is not less than \code{maxdist} in edit distance from at least one of the non-\code{NA} sequences in \code{asvs}: \code{NA_character_}}}}}
}
