% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara-package.R
\docType{data}
\name{quality_profile}
\alias{quality_profile}
\title{Sample quality scores}
\format{
A \code{data.frame} with columns \code{seq_id}, \code{quality},
and \code{n}, giving the number of times each wuality score appears in each
read.
}
\usage{
quality_profile
}
\description{
These are based on CCS quality scores from the first 1000 reads of long
amplicon PacBio RSII samples in
\href{https://www.authorea.com/doi/full/10.22541/au.160340221.18389016/v1}{Furneaux et al. 2020}.
}
\keyword{datasets}
