% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{seqhash}
\alias{seqhash}
\alias{seqhash.character}
\alias{seqhash.XStringSet}
\title{Individually hash biological sequences}
\usage{
seqhash(seq, algo = "xxhash32", len = NA, preserve_na = TRUE)

\method{seqhash}{character}(seq, algo = "xxhash32", len = NA, preserve_na = TRUE)

\method{seqhash}{XStringSet}(seq, algo = "xxhash32", len = NA, preserve_na = TRUE)
}
\arguments{
\item{seq}{(\code{character} or \code{\link[Biostrings]{XStringSet-class}})
the sequences to hash.}

\item{algo}{(\code{character}) a hash algorithm supported by
\code{\link[digest]{digest}}. default: "xxhash32"}

\item{len}{(\code{integer}) number of characters to keep from each hash
string. \code{NA} (the default) to keep all characters.}

\item{preserve_na}{(\code{logical}) If \code{TRUE}, \code{NA} values in
\code{seq} are preserved as \code{NA} in the output.  If
\code{FALSE}, then \code{NA} is passed to
\code{\link[digest]{digest}}, which results in a valid hash.}
}
\value{
a \code{character} vector of the same length as \code{seq},
    with the hashed sequences.
}
\description{
Individually hash biological sequences
}
