% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{summarize_sread}
\alias{summarize_sread}
\title{Summarize one or more reads as a \code{\link[tibble]{tibble}}.}
\usage{
summarize_sread(sread, ..., max_ee)
}
\arguments{
\item{sread}{(\code{\link[ShortRead:ShortReadQ-class]{ShortReadQ}} object or
list of such objects) as produced by
\code{\link[ShortRead]{readFastq}}.}

\item{...}{(any vector) additional columns to add to the output
\code{\link[tibble]{tibble}}.  These should have length 1 or, if
\code{sread} is a named list of
\code{\link[ShortRead:ShortReadQ-class]{ShortReadQ}}, the length of
\code{sread}.  Recycling behavior is as in
\code{\link[tibble]{tibble}}. Avoid the name "\code{name}" if
\code{sread} is a named \code{list}}

\item{max_ee}{(\code{numeric}) filter out reads with expected error greater
than \code{max_ee}.  Default: Inf (no filtering)}
}
\value{
a \code{\link[tibble]{tibble}} with columns: \describe{
        \item{\code{$seq_id} (\code{character})}{sequence IDs, typically from
            the input fastq}
        \item{\code{$seq} (\code{character})}{the actual sequence reads}
        \item{\code{$name} (\code{character})}{the name of the source
            \code{\link[ShortRead:ShortReadQ-class]{ShortReadQ}}, if
            \code{sread} was a list. Otherwise absent.}
        \item{\code{...}}{any other arguments passed to summarize_sread.}}
    If the input was empty, not a valid
    \code{\link[ShortRead:ShortReadQ-class]{ShortReadQ}}, or no sequences
    passed filtering, a \code{\link[tibble]{tibble}} with zero rows is
    returned.
}
\description{
Summarize one or more reads as a \code{\link[tibble]{tibble}}.
}
