% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara-package.R
\docType{data}
\name{tzara_sample}
\alias{tzara_sample}
\title{Sample amplicon sequences}
\format{
A \code{\link[Biostrings]{DNAStringSet}} with three sequences.
}
\usage{
tzara_sample
}
\description{
These are the three most abundant ASVs from
\href{https://www.authorea.com/doi/full/10.22541/au.160340221.18389016/v1}{Furneaux et al. 2020}.
They are derived from PacBio RSII metabarcoding of soil samples from
central Benin, West Africa.
The three sequences are from the fungal genera \emph{Penicillium}, \emph{Agaricus}
(included as a control), and \emph{Inocybe}.
}
\keyword{datasets}
