% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara.R
\name{cluster_consensus}
\alias{cluster_consensus}
\alias{cluster_consensus.character}
\alias{cluster_consensus.XStringSet}
\title{Calculate consensus of a cluster of sequences.}
\usage{
cluster_consensus(seq, ..., ncpus = 1, simplify = TRUE)

\method{cluster_consensus}{character}(seq, names = names(seq),
  DNA2RNA = TRUE, ..., ncpus = 1, simplify = TRUE)

\method{cluster_consensus}{XStringSet}(seq, ..., ncpus = 1,
  simplify = TRUE)
}
\arguments{
\item{seq}{(\code{character} vector or \code{\link[Biostrings]{XStringSet}})
The sequences to calculate a consensus for.}

\item{...}{passed to methods}

\item{ncpus}{(\code{integer}) Number of CPUs to use.}

\item{simplify}{(\code{logical}) If \code{TRUE}, return an object of the same
type as \code{seq} containing a single sequence representing the consensus.
If \code{FALSE}, an object of the same type as \code{seq} representing the
consensus sequence for reads which were included in the consensus, or
\code{NA_character_} for reads which were initially \code{NA} or which were
removed from the consensus alignment as outliers.  For the
\code{\link[Biostrings]{XStringSet}} method, which does not allow \code{NA}
entries, these elements are missing from the set (this can be deduced by
the names).}

\item{names}{(\code{character}) If \code{seq} is a \code{character} vector,
names for the sequences.}

\item{DNA2RNA}{(logical) whether to convert \code{seq} from DNA to RNA, and use (calculated) RNA secondary structure in alignments.}
}
\value{
an \code{\link[Biostrings]{XStringSet}} representing the consensus
sequence.
}
\description{
This algorithm assumes that the sequences "should be" identical except for
amplification and sequencing errors.  Its main purpose is to calculate a
consensus sequence for an amplicon that is too long to use in DADA2 directly,
but which has been clustered based on sequence variant identity in one
subregion.
}
\details{
The sequences are first aligned using
\code{\link[DECIPHER]{AlignSeqs}}. Sequences which are "outliers" in the
alignment are then removed by
\code{\link[odseq]{odseq}}. If the input sequences were clustered based on
DADA2 sequence variants of a variable region, and the sequences were
appropriately quality filtered prior to running \code{\link[dada2]{dada}},
then outliers should mostly be chimeras.

After outlier removal, sites with greater than 50\% gaps are removed, and
the most frequent letter (ignoring gaps) is chosen at all other sites. If no
letter has greater than 50\% representation at a position, then an IUPAC
ambiguous base representing at least 50\% of the reads at that position is
chosen for nucleotide sequences, or \code{"X"} for amino acids.
}
