% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzara-package.R
\docType{data}
\name{tzara_sample}
\alias{tzara_sample}
\title{Sample amplicon sequences}
\format{
A \code{\link[Biostrings]{DNAStringSet}} with three sequences.
}
\usage{
tzara_sample
}
\description{
These are the three most common ASVs from Furneaux et al. 2020 (submitted).
They are derived from PacBio RSII metabarcoding of soil samples from
central Benin, West Africa.
The three sequences are from the fungal genera _Penicillium_, _Agaricus_
(used as a control sample), and _Inocybe_.
}
\keyword{datasets}
